feat(interface): add  ValidatorFieldsInterface
Created the ValidatorFieldsInterface to represent the structure of an object with fields' errors
and validated data.The interface contains the "errors" property, which holds the validation errors
for each field represented as a dictionary. It also includes the "validatedData" property, which
stores the data that passed the validation. Implemented the "validate" method that takes data of
type T as input and returns a boolean indicating whether the data is valid or not.