feat: add Search Parameters and Search Repository Interface
Added a SearchParams class that encapsulates search parameters like page, perPage, sort, sortDir,
and filter for use in repositories. Implemented validation and default values for each parameter to
ensure proper initialization and prevent invalid values. Created a SearchRepositoryInterface that
extends the RepositoryInterface with an additional search method for performing searches using the
provided search parameters. This interface is intended to be implemented by repositories that
support search functionality.