feat: add UserInMemoryRepository
Created the UserInMemoryRepository class that extends InMemorySearchRepository and implements the
UserRepository interface. Added the findByEmail method to retrieve a user entity based on the
provided email. If the entity is not found, a NotFoundError is thrown. Added the checkEmailExists
method to check if a user entity with the provided email already exists. If an entity is found, a
ConflictError is thrown.