refactor: refactor SearchParams and SearchResult classes and their interfaces
Updated the SearchProps and SearchResultProps types to be more flexible by allowing the Entity type
to have any unknown type. Modified the SearchParams class to accept a generic type parameter Filter
and ensure that the _filter property is of type Filter | null. Adjusted the
SearchRepositoryInterface to reflect the changes in the SearchProps and SearchResultProps types.
Added a sortableFields property to the SearchRepositoryInterface to indicate the fields that can be
used for sorting.