test: add unit tests for UserInMemoryRepository methods
Added unit tests for the UserInMemoryRepository class. Checked that an error is thrown when an
email is not found. Checked that a user is found by email when it exists. Checked that an error is
thrown when an email already exists. Checked that no error is thrown when an email does not exist.
Checked that items are not filtered when the filter object is null. Checked that items are filtered
by the name field using the provided filter parameter. Checked that items are sorted by createdAt
when the sort parameter is null. Checked that items are sorted by the name field in ascending and
descending order.