feat: implement SignupService for User Registration
Created the SignupService namespace to handle user registration. Defined the Input and Output types
for the service. Implemented the Create class within the namespace, responsible for user
registration. The Create class constructor takes a userRepository and hashProvider as dependencies.
The executeSignup method validates input data, throwing a BadRequestError for incomplete data. The
method checks for existing email addresses using the userRepository.emailExists method. The password
is securely hashed using the provided hashProvider. A new UserEntity instance is created and
persisted using the userRepository.create method. The method returns the serialized representation
of the created user.