test: add Unit Tests for DetailUserService
Created unit tests for the DetailUserService to ensure its functionality. Implemented test cases for
the findOne method: Tested for throwing an error when an invalid ID is provided. Tested for
retrieving user details when a valid ID is provided. Utilized spies to verify repository method
calls and assertions to ensure correct behavior.