feat: implement User Profile Detail Service
Created the DetailUserService module to handle user profile details. Defined the Input type as an
object with an id property.Defined the Output type to match the structure of the OutputUser DTO.
Implemented the UserProfile class with a constructor that injects the userRepository dependency.
Implemented the findOne method to fetch user details using the userRepository and return them as a
serialized JSON object.