feat: implement PaginatedItemsMapper class
Added the PaginatedItemsMapper class. Created a static method toPaginatedItems which takes an array
of items and a SearchResult object as parameters. The method converts the SearchResult object into a
PaginatedItems object by extracting relevant properties. This class is used to map a SearchResult
object from the repository to a more general PaginatedItems format for use in the application.