feat: add UserProfileModification service
Added the UserProfileModification service in the ModifyUserService module. Implemented the update
method to handle updating a user's profile. Checked if the name field is provided in the input,
throwing a BadRequestError if not. Used the findOne method from the repository to retrieve the user
entity. Invoked the updateUserName method of the entity to update the user's name. Finally, called
the update method of the repository and returned the updated user details using the
UserOutputMapper.