feat: add SignInService for user authentication
Added the SignIn use case implementation for user authentication. The SignIn class implements the
ApplicationAction interface to provide the executeSignIn method. The executeSignIn method
authenticates users by checking the provided email and password. It throws a BadRequestError if
email or password is missing. It retrieves the user entity from the repository by email and compares
the provided password with the hashed password using the hashProvider. If the passwords match, the
user is authenticated, and the user's details are returned using UserOutputMapper. If the passwords
don't match, it throws an InvalidCredentialsError. Also, it added proper parameter validation and
error handling to ensure the correctness of the authentication process.