feat: add UserModelMapper for Entity Conversion
Implemented UserModelMapper class to convert a Prisma User model into a UserEntity. Created a static
method toEntity which takes a User model as input and constructs a UserEntity instance using the
provided data. Handled the conversion process and ensured that any errors during entity creation are
caught and rethrown as a ClassValidatorFields error for consistent error handling.