test: add Integration Tests for UserModelMapper
Added integration tests for the UserModelMapper class. Used PrismaClient to interact with the
database for testing. Set up setupPrismaTest to initialize the test environment for Prisma. Defined
props to hold test data for user properties. Implemented beforeAll and afterAll hooks to connect and
disconnect from the Prisma service. Implemented beforeEach to clear the user table before each test.
Created tests to validate error handling when the user model is invalid and to test the conversion
of a user model to a user entity using UserModelMapper.toEntity. Used expect assertions to validate
the expected behaviors of the mapper functions and the resulting entities.