feat: implement CollectionPresenter with pagination metadata
Added the CollectionPresenter class.  Created a protected property paginationPresenter of type
PaginationPresenter to manage pagination metadata. In the constructor, initialized the
paginationPresenter with the provided properties. Decorated the meta getter with @Expose and
configured it to be named 'meta' in the serialized object. Defined an abstract data getter in the
CollectionPresenter class, which should be implemented by subclasses to provide the collection data.