feat: implement WrapperDataInterceptor
Added the WrapperDataInterceptor class responsible for intercepting requests in NestJS. Implemented
the NestInterceptor interface. Utilized the intercept method to process the request context and call
handler. Used the map operator from rxjs to modify the response body. Modified the response
structure to wrap the body in a 'data' property if not already wrapped. Checked for the presence of
a 'meta' key in the body and returned it unchanged if present.